Major rework to how blade deals damage
Every attack now holds a temporary state

This state determines at which frame the enemy is going to get hurt.
The goal was to write this flexible enough to integrate range checks into individual damage frames. For now it is a bare-bones implementation of delayed and multi-instance damage dealing mechanic.

(Hurting the same enemy multiple times in the same attack feels really good btw.)

replaced modifier to damage for hitting multiple enemies in one swing to hurting the same enemy multiple times (Individual damage numbers are lowered so this parameter is not needed anymore.)

Disabled particle spawning until I can figure out a sane equation for the number of particles to spawn.

Also added new sounds for attacking and hitting enemies